Scrapping/Australia_GroceriesScraper/requirements.txt
pymongo==4.7.0
selenium==4.11.2
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
configparser==5.3.0
